requirements.txt
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.10
gradio>=3.0.0
termcolor>=2.0.0
python-docx>=0.8.11
beautifulsoup4>=4.11.1